keyvault: add script to retrieve Key Vault secrets via Azure CLI
Oftentimes, it is useful to access secrets stored in an Azure Key
Vault during VM provisioning. For example, a new VM may need to
have sensitive files such as SSH keys or SSL certificates
stored in disk. One option is to retrieve stored secrets from a
Key Vault when the VM is first launched.

This script abstracts the complexity of accessing an existing
Key Vault. For each secret in the Vault, a text file with its
contents being the secret value will be created in the user's
home directory i.e. /home/azureuser/mysshkey.txt. The files
will be accessible only to the current user.

Cloud-init or post-launch scripts may handle these files as they
see fit.

Usage:
        get_azure_key_vault_secret.py <1> <2*>
        <1>: Name of Key Vault from which to retrieve secrets
        <2*>: One or more names of secrets stored in the Key Vault

Example:
        get_azure_key_vault_secret.py MyKeyVault MySSHKeyName MySSLCertName

This script assumes that the Azure Linux VM has an identity
assigned and the Key Vault has been configured to provide
get/decrypt access to the VM's identity.