keyvault: add exception handling for non-existing secret in vault
Handle the cases where the provided secret name does not exist
in a Key Vault and when a key vault is inaccessible to the VM
executing the script. The latter can happen if
'az keyvault set-policy' has not been run either on the host
that created the VM or in an authenticated az cli session.